pyproject.toml
[project]
name = "whisper-ime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai-whisper",
    "sounddevice",
    "soundfile",
    "pynput",
]
